refactor: delete dead code and unused dependencies
Removes:
- Dead floating ShortcutsOverlay component + test (A1)
- Unused event-view.adapter.ts + types cascade (A2)
- Unused recurrence.util.ts (A2)
- Unused npm dependencies: dayjs, react-textarea-autosize, tailwindcss, typescript-plugin-css-modules (A3)
- skipOnboarding field and all references (A5)
- Export qualifiers from unused schemas (A6 partial)

Moves ShortcutOverlaySection type to new shortcuts-overlay.types.ts file to
resolve the ARIA role conflict between dead and live components.

Type-check clean. All references verified removed via grep.
