+++ packages/web/package.json
@@ -21,7 +21,6 @@
     "bson": "^7.2.0",
     "classnames": "^2.3.1",
     "clsx": "^1.2.1",
-    "dayjs": "^1.10.7",
     "dexie": "^4.2.1",
     "dompurify": "^3.4.13",
     "fast-deep-equal": "^3.1.3",
@@ -30,7 +29,6 @@
     "react-datepicker": "^4.2.1",
     "react-dom": "^18.1.0",
     "react-select": "^5.8.0",
-    "react-textarea-autosize": "^8.3.3",
     "react-toastify": "^9.1.3",
     "rrule": "^2.7.2",
     "supertokens-web-js": "^0.16.0",
@@ -57,9 +55,7 @@
     "jsdom": "^26.1.0",
     "msw": "^1.0.1",
     "postcss": "^8.5.6",
-    "tailwindcss": "^4.1.14",
-    "typescript": "^7.0.2",
-    "typescript-plugin-css-modules": "^5.0.1"
+    "typescript": "^7.0.2"
   },
   "scripts": {
     "build": "bun run build.ts",
