--- packages/web/src/components/Shortcuts/ShortcutOverlay/ShortcutsOverlay.tsx
@@ -1,54 +1,0 @@
-import classNames from "classnames";
-import { ZIndex } from "@web/common/constants/web.constants";
-import { type Shortcut } from "@web/shortcuts/global.shortcut.types";
-import { ShortcutSection } from "./ShortcutSection";
-
-export interface ShortcutOverlaySection {
-  id?: string;
-  title: string;
-  shortcuts: Shortcut[];
-}
-
-interface ShortcutsOverlayProps {
-  sections: ShortcutOverlaySection[];
-  heading?: string;
-  ariaLabel?: string;
-  className?: string;
-}
-
-export const ShortcutsOverlay = ({
-  sections,
-  heading = "Shortcuts",
-  ariaLabel = "Shortcut overlay",
-  className = "",
-}: ShortcutsOverlayProps) => {
-  const visibleSections = sections.filter(
-    (section) => section.shortcuts.length > 0,
-  );
-
-  if (!visibleSections.length) return null;
-
-  return (
-    <aside
-      aria-label={ariaLabel}
-      className={classNames(
-        "fixed top-24 left-3 border-border bg-surface",
-        "border p-3 shadow-lg backdrop-blur-sm md:block",
-        `hidden w-60 rounded-lg ${className}`,
-      )}
-      style={{ zIndex: ZIndex.MAX }}
-    >
-      {heading && (
-        <div className="mb-2 font-medium text-text text-xs">{heading}</div>
-      )}
-      {visibleSections.map((section, index) => (
-        <ShortcutSection
-          key={section.id ?? section.title}
-          isFirst={index === 0}
-          title={section.title}
-          shortcuts={section.shortcuts}
-        />
-      ))}
-    </aside>
-  );
-};
