+++ packages/web/src/shortcuts/shortcuts-overlay.types.ts
@@ -1,0 +1,7 @@
+import { type Shortcut } from "@web/shortcuts/global.shortcut.types";
+
+export interface ShortcutOverlaySection {
+  id?: string;
+  title: string;
+  shortcuts: Shortcut[];
+}
